pyproject.toml
[project]
name = "my-project"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12,<3.14"
dependencies = [
    "anthropic>=0.86.0",
    "dash>=4.0.0",
    "dash-bootstrap-components>=2.0.4",
    "edgartools>=5.25.1",
    "fastapi>=0.135.2",
    "kaleido>=1.2.0",
    "lightgbm>=4.6.0",
    "lseg-data>=2.1.0",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "pandasdmx>=1.10.0",
    "pandera>=0.30.0",
    "pmdarima>=2.1.1",
    "prophet>=1.3.0",
    "pyarrow>=23.0.1",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "requests-cache>=1.3.1",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.1",
    "shap>=0.46.0",
    "statsmodels>=0.14.6",
    "tqdm>=4.67.3",
    "uvicorn>=0.42.0",
    "wbgapi>=1.0.14",
    "weasyprint>=68.1",
]

[tool.uv]
# Windows is not a supported deployment target for this project.
# edgartools (pydantic>=2, httpxthrottlecache>=0.3.5) and lseg-data/pandasdmx
# have conflicting transitive dependencies only under Windows-specific markers.
# Restricting resolution to macOS/Linux resolves all conflicts.
environments = ["sys_platform != 'win32'"]
# Dependency overrides to resolve edgartools vs pandasdmx/lseg-data conflicts:
# - pandasdmx declares pydantic<2, edgartools requires pydantic>=2 → force v2
#   (pandasdmx breaks at runtime with pydantic v2, but is only used in existing
#    oecd.py / test_oecd.py which mock the library at the module level)
# - lseg-data declares httpx<0.28, edgartools needs httpx>=0.28.1 → force >=0.28.1
#   (lseg-data 2.1.0 is verified to work at runtime with httpx 0.28.1)
override-dependencies = ["pydantic>=2.0.0", "httpx>=0.28.1"]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "jupyterlab>=4.5.6",
    "pytest>=9.0.2",
    "ruff>=0.15.6",
]

